requirements.txt
# 核心依赖
python-dotenv>=0.19.0
loguru>=0.5.3

# Web相关
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.26.0

# 图片处理
Pillow>=8.0.0

# 工具
cloudscraper>=1.2.71
urllib3>=2.0.7
